requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
motor>=3.1.1
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
email-validator>=1.3.1
httpx>=0.24.0
